fix(cfd): correct a false deferral, a wrong rationale, and two misplaced files
Follow-up to 5d5c38732. Acting on the lesson that change 4's design named the
wrong blocker for the wall heat flux, every deferral making a factual claim
about the tree was tested against it. Two were false.

**Nav attitude (a) is not "feature-sized".** `fix-navigation-filter-correctness`
D4 chose resolution (b) — retain the attitude error rather than inject it —
solely because (a) was judged too large: "the engine needs an attitude
representation (quaternion or DCM)". It already has one. `Quaternion`,
`Quaternion32`, `Quaternion64` ship in `deep_causality_num_complex`, which
`deep_causality_cfd` already depends on, carrying exactly the ESKF operations:
`from_axis_angle` for injecting δψ, `to_rotation_matrix` for the −[f]× coupling,
`normalize` for integration drift, `from_euler_angles` and `slerp`.

So (a) is a field on `ReentryNavEngine` (which today holds position, velocity,
filter, tau_offset, elapsed and no attitude) plus two call sites. (a) now lands
in that change and (b) is dropped: (b) existed only to avoid (a)'s cost and
carried a real price — an attitude block that is no longer an error about the
current nominal, a documented departure from textbook ESKF bookkeeping that
would later have to be undone. (a) also fixes the original defect properly,
since `reset_navigation` may then zero the block *because the correction was
actually injected*, which is the invariant the spec states.

**`collect_constrained_edges` has no known consumer.** Change 4's D4 reversed a
removal recommendation citing `aperture-resolved-noslip` as the capability that
would supply zone constraints. That capability is already implemented, is the
*default* (`NoSlipConstraint::new(.., true)`), and derives its constraints
through `CutCellRegistry::cut_face_constraints` in `no_slip.rs` — not through
the hook. The justification was wrong. Corrected at both docstrings rather than
left in place: the hook is kept because union composition is idempotent and it
is behaviourally covered, but it is an extension point without a claimed user
and should be removed if none appears — not defended by citing a capability
that solved the problem elsewhere.

Three deferrals were confirmed TRUE and stand: item 12 (negative pressure
reaches the flux — `marcher_2d.rs:135` floors `p` only for the wave speed),
item 13 (QTT constructors validate nothing), item 14 (mask not clamped). RAMC's
"the composition it needs is already computed" is also true —
`air_n2_mole_fraction` / `air_o2_mole_fraction` are shipped kernels already
called at `qtt_ramc_stagline/main.rs:197-198` — so that follow-up is small and
should be raised when change 1 lands rather than carried as a disclaimer.

Housekeeping, both AGENTS.md conformance:

- `wall_heat_flux` tests moved into `surface_force_tests.rs`. AGENTS.md requires
  a test file mirror the *source* file name; `wall_heat_flux` lives in
  `surface_force.rs`, so a separate `wall_heat_flux_tests.rs` was wrong. All six
  tests and both helpers verified present before the old file was removed.
