+++ deep_causality_cfd/tests/solvers/dec/mod.rs
@@ -45,6 +45,5 @@ pub mod taylor_green_3d_tests;
 pub mod uncertain_boundary_source_tests;
 #[cfg(all(test, feature = "std"))]
 pub mod uncertain_inflow_tests;
-pub mod wall_heat_flux_tests;
 #[cfg(test)]
 pub mod wrappers_tests;
